настроил политику для аттачментсов и добавил в искоючения проверки если контроллер devise

+++ app/controllers/product_attachments_controller.rb
@@ -1,13 +1,21 @@
 class ProductAttachmentsController < ApplicationController
+
+
   def edit
     @product_attachment = ProductAttachment.find(params[:id])
+    authorize @product_attachment
   end
 
   def destroy
     @product_attachment = ProductAttachment.find(params[:id])
+    authorize @product_attachment
     if @product_attachment.destroy
-      flash[:success] = "Изображение удалено!"
+      flash[:success] = 'Изображение удалено!'
       redirect_to :back
     end
   end
+
+  private
+
+
 end
